+++ ui/app/utils/properties/job-client-status.js
@@ -1,4 +1,5 @@
 import { computed } from '@ember/object';
+import matchGlob from '../match-glob';
 
 const STATUS = [
   'queued',
@@ -26,7 +27,9 @@ export default function jobClientStatus(nodesKey, jobKey) {
 
       // Filter nodes by the datacenters defined in the job.
       const filteredNodes = nodes.filter((n) => {
-        return job.datacenters.indexOf(n.datacenter) >= 0;
+        return job.datacenters.find((dc) => {
+          return !!matchGlob(dc, n.datacenter);
+        });
       });
 
       if (job.status === 'pending') {
