+++ ui/tests/acceptance/job-detail-test.js
@@ -30,6 +30,18 @@ moduleForJobWithClientStatus(
     })
 );
 
+moduleForJobWithClientStatus(
+  'Acceptance | job detail with client status (system with wildcard dc)',
+  () =>
+    server.create('job', {
+      id: 'system-wildcard-dc',
+      status: 'running',
+      datacenters: ['canada-*-1'],
+      type: 'system',
+      createAllocations: false,
+    })
+);
+
 moduleForJob('Acceptance | job detail (sysbatch)', 'allocations', () =>
   server.create('job', { type: 'sysbatch', shallow: true })
 );
@@ -52,6 +64,20 @@ moduleForJobWithClientStatus(
     return server.create('job', {
       status: 'running',
       datacenters: ['dc1'],
+      type: 'sysbatch',
+      namespaceId: namespace.name,
+      createAllocations: false,
+    });
+  }
+);
+
+moduleForJobWithClientStatus(
+  'Acceptance | job detail with client status (sysbatch with namespace and wildcard dc)',
+  () => {
+    const namespace = server.create('namespace', { id: 'test' });
+    return server.create('job', {
+      status: 'running',
+      datacenters: ['*'],
       type: 'sysbatch',
       namespaceId: namespace.name,
       createAllocations: false,
@@ -89,6 +115,20 @@ moduleForJobWithClientStatus(
       shallow: true,
       namespaceId: namespace.name,
       datacenters: ['dc1'],
+    });
+    return server.db.jobs.where({ parentId: parent.id })[0];
+  }
+);
+
+moduleForJobWithClientStatus(
+  'Acceptance | job detail with client status (sysbatch child with namespace and wildcard dc)',
+  () => {
+    const namespace = server.create('namespace', { id: 'test' });
+    const parent = server.create('job', 'periodicSysbatch', {
+      childrenCount: 1,
+      shallow: true,
+      namespaceId: namespace.name,
+      datacenters: ['*'],
     });
     return server.db.jobs.where({ parentId: parent.id })[0];
   }
